requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
aiofiles==23.2.1
pydantic[email]==2.5.0
paypalrestsdk==1.13.1
